feat(live-tv): a durable guide, programme identity, and a fortnight horizon
Three changes that only work together, which is why they are one commit:
a 14-day guide cannot be fetched in one refresh, accumulating it across
refreshes is only safe once a restart does not start from nothing, and a
series recording rule needs both the horizon and an identity to match on.

The guide survives a restart
  The owner's copy is written to `<runtime cache>/live-tv/guide.json`,
  outside the state mutex and always tmp-then-rename. A copy loaded at
  startup keeps the age it already had: `CachedGuide::age_offset` exists
  because `Instant` has no epoch, so a machine up for less time than the
  guide's age has no instant to subtract to, and without the offset that
  copy would claim to be new. The refresh loop now also wakes on the three
  things that change its answer — a lineup arriving cold, a settings save,
  and a serving-fence transition — rather than only on a twenty-minute
  timer, and reads the lineup once itself when nothing has read it yet.
  Every guide document carries `next_refresh_at`, an `unavailable` one
  included, so a client polls on the owner's clock.

Programmes carry identity
  `series_id`, `programme_id` and `is_new` on `LiveTvProgramme`, from
  HDHomeRun's `SeriesID`/`ProgramID` — spelled with explicit serde renames,
  because `rename_all = "PascalCase"` looks for `SeriesId` and would have
  parsed nothing — and from an XMLTV `dd_progid` and `<new/>`. A source
  that supplies neither parses exactly as it did before.

The horizon reaches a fortnight
  `MAX_GUIDE_HOURS` 72 → 336 (and the request bound, and the settings
  field's type), the per-channel cap 200 → 800 because ~670 rows is what a
  fortnight of half-hour slots costs, per-channel extension that
  accumulates against the persisted cache within its 64-request budget,
  and one revalidated day per tick cycling the horizon so a provider's
  schedule change reaches the cache within ~5 hours rather than never.

One behaviour change worth calling out: an overlap now shortens the
earlier row's end instead of moving the later row's start, and a row
wholly inside another is dropped. A start is about to become the DVR's
identity for an airing, so a normalisation that could move one would
silently re-key every scheduled recording on the channel.

+++ crates/plurxd/src/http/system.rs
@@ -1554,7 +1554,7 @@ pub struct SettingsDto {
     pub live_tv_transition_drain_before: i64,
     pub live_tv_guide_source: String,
     pub live_tv_xmltv_url: String,
-    pub live_tv_guide_hours: u8,
+    pub live_tv_guide_hours: u16,
     pub tmdb_configured: bool,
     /// The stored TMDB key itself. This endpoint is admin-only and the key is
     /// low-sensitivity (read-only metadata), so the admin who set it can see
@@ -2169,7 +2169,7 @@ pub struct UpdateSettings {
     /// enabled, because turning a read-only feed on must not drain viewers.
     pub live_tv_guide_source: Option<String>,
     pub live_tv_xmltv_url: Option<String>,
-    pub live_tv_guide_hours: Option<u8>,
+    pub live_tv_guide_hours: Option<u16>,
     /// Explicit admin attestation, never an automatic timeout override.
     pub live_tv_fenced_owner: Option<LiveTvFencedOwner>,
     /// Set the TMDB API key. Empty string clears it. Absent leaves it as-is.
